[CHANGE] Velocity Component is dead, RIP

+++ sexy_engine/include/ecs/Components/PosComponent.hpp
@@ -11,8 +11,8 @@ namespace Sex {
     struct PosComponent : public IComponent {
         PosComponent(int xx, int yy) : x(xx), y(yy) {}
 
-        unsigned int x;
-        unsigned int y;
+        double x;
+        double y;
     };
 }
